+++ examples/src/popups/ExamplePopup.ts
@@ -1,5 +1,5 @@
 import { buttonStyle, GREEN } from '@/utils/Constants';
-import { Container, Popup } from 'dill-pixel';
+import { Container, forceFocus, Popup, registerFocusables } from 'dill-pixel';
 import { gsap } from 'gsap';
 import { Point, Sprite, Text } from 'pixi.js';
 
@@ -14,12 +14,6 @@ export default class ExamplePopup extends Popup {
 
   init(size: Point) {
     super.init(size);
-
-    this.alpha = 0;
-
-    if (this.blackout) {
-      this.blackout.alpha = 0;
-    }
 
     this._bg = this.add.coloredSprite({
       color: GREEN,
@@ -47,23 +41,25 @@ export default class ExamplePopup extends Popup {
     this._button.add.text({ value: 'Close me', anchor: 0.5, style: { ...buttonStyle, fontSize: 18 } });
     this._button.eventMode = 'static';
     this._button.cursor = 'pointer';
+    this._button.focusable = true;
 
-    this._button.on('pointerdown', (e) => {
-      this.hide();
-    });
+    registerFocusables(this._button);
+
+    this._button.on('pointerdown', this.close);
+    this._button.onFocusActivated = this.close;
   }
 
   async animateIn(callback: () => void): Promise<void> {
-    this.y = 10;
     if (this.blackout) {
-      await gsap.to(this.blackout, { alpha: 1, visible: true, duration: 0.3 });
+      gsap.fromTo(this.blackout, { alpha: 0 }, { alpha: 1, visible: true, duration: 0.3 });
     }
-    await gsap.to(this, { alpha: 1, y: 0, visible: true, ease: 'sine.out', duration: 0.4 });
+    await gsap.fromTo(this, { alpha: 0, y: 20 }, { alpha: 1, y: 0, visible: true, ease: 'sine.out', duration: 0.4 });
+    forceFocus(this._button);
     callback();
   }
 
   async animateOut(callback: () => void): Promise<void> {
-    gsap.to(this, { alpha: 0, y: -5, ease: 'sine.in', duration: 0.3 });
+    gsap.to(this, { alpha: 0, y: -10, ease: 'sine.in', duration: 0.3 });
     if (this.blackout) {
       await gsap.to(this.blackout, { alpha: 0, duration: 0.2, delay: 0.1 });
     }